requirements.txt
streamlit>=1.31.0
pandoc>=2.3
python-magic>=0.4.27